refactor(security): Move certificate trust operation
